refactor: improve courses & categories management
- Enforce uniqueness for courseName and categoryName

- Implement full CRUD operations

- Add personalized messages for each controller

+++ src/main/java/com/emmlg/ForoAluraHub/curso/modelo/Course.java
@@ -14,7 +14,7 @@ public class Course {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "id_course")
-    private long courseId;
+    private Integer courseId;
 
     @Column(name = "course_title")
     private String courseName;
